requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
numpy==1.16.4
tensorflow-gpu==1.14.0
six==1.12.0
requests==2.22.0
